fix: move ctype utilities
