+++ src/mappings/cTypes/cTypeHasher.ts
