+++ src/mappings/cTypes/extractCTypeDefinition.ts
